chore: refactor bridge into separate modules
