+++ src/bridge/deposit.ts
@@ -1,149 +1,13 @@
 import BigNumber from 'bignumber.js';
 import { ethers } from 'ethers';
 
-import { assetUnitsToDecimal, decimalToPip, multiplyPips } from '#pipmath';
+import { multiplyPips } from '#pipmath';
 
 import { getExchangeAddressAndChainFromApi } from '#client/rest/public';
-import { ExchangeLayerZeroAdapter_v2__factory } from '#typechain-types/factories/ExchangeLayerZeroAdapter_v2__factory';
-import { KumaStargateForwarder_v1__factory } from '#typechain-types/factories/KumaStargateForwarder_v1__factory';
 import { IOFT__factory } from '#typechain-types/index';
 import { StargateV2Target } from '#types/enums/request';
 
-import {
-  StargateV2Config,
-  StargateV2ConfigByLayerZeroEndpointId,
-} from './config';
-
-import type {
-  DecodedStargateV2Payload,
-  EncodedStargateV2Payload,
-} from '#types/bridge';
-import type { BridgeTarget } from '#types/enums/request';
-
-export const StargateV2MainnetLayerZeroEndpointIds = Object.values(
-  StargateV2ConfigByLayerZeroEndpointId.mainnet,
-).map((value) => {
-  return value.layerZeroEndpointId;
-});
-
-export type StargateV2LayerZeroEndpointIdsMainnet =
-  (typeof StargateV2MainnetLayerZeroEndpointIds)[number];
-
-export function isStargateV2MainnetLayerZeroEndpointId(
-  layerZeroEndpointId: number,
-): layerZeroEndpointId is StargateV2LayerZeroEndpointIdsMainnet {
-  return StargateV2MainnetLayerZeroEndpointIds.includes(
-    layerZeroEndpointId as StargateV2LayerZeroEndpointIdsMainnet,
-  );
-}
-
-export const StargateV2TestnetLayerZeroEndpointIds = Object.values(
-  StargateV2ConfigByLayerZeroEndpointId.testnet,
-).map((value) => {
-  return value.layerZeroEndpointId;
-});
-
-export type StargateV2LayerZeroEndpointIdsTestnet =
-  (typeof StargateV2TestnetLayerZeroEndpointIds)[number];
-
-export function isStargateV2TestnetLayerZeroEndpointId(
-  layerZeroEndpointId: number,
-): layerZeroEndpointId is StargateV2LayerZeroEndpointIdsTestnet {
-  return StargateV2TestnetLayerZeroEndpointIds.includes(
-    layerZeroEndpointId as StargateV2LayerZeroEndpointIdsTestnet,
-  );
-}
-
-/**
- * Get a stargate config with strict typing to allow narrowing on the
- * `config.supported` boolean
- *
- * @example
- * ```typescript
- * const config = getStargateTargetConfig(StargateTarget.STARGATE_ARBITRUM, true);
- * ```
- */
-export function getStargateV2TargetConfig<
-  T extends StargateV2Target,
-  S extends true | false,
->(stargateTarget: T, sandbox: S) {
-  const targetConfig =
-    sandbox ?
-      StargateV2Config.testnet[stargateTarget]
-    : StargateV2Config.mainnet[stargateTarget];
-  if (!targetConfig) {
-    throw new Error(
-      `No config found for ${stargateTarget} ${sandbox ? 'testnet' : 'mainnet'}`,
-    );
-  }
-
-  return targetConfig;
-}
-
-export function stargateV2TargetForLayerZeroEndpointId(
-  layerZeroEndpointId: number,
-  sandbox: boolean,
-) {
-  if (sandbox && isStargateV2TestnetLayerZeroEndpointId(layerZeroEndpointId)) {
-    return StargateV2ConfigByLayerZeroEndpointId.testnet[layerZeroEndpointId]
-      .target;
-  }
-  if (!sandbox && isStargateV2MainnetLayerZeroEndpointId(layerZeroEndpointId)) {
-    return StargateV2ConfigByLayerZeroEndpointId.mainnet[layerZeroEndpointId]
-      .target;
-  }
-
-  return null;
-}
-
-/**
- * Decode an ABI-encoded hex string representing Stargate V2 withdrawal parameters
- */
-export function decodeStargateV2Payload(
-  payload: EncodedStargateV2Payload,
-): DecodedStargateV2Payload {
-  return {
-    layerZeroEndpointId: parseInt(
-      ethers.AbiCoder.defaultAbiCoder().decode(['uint32'], payload)[0],
-      10,
-    ),
-  };
-}
-
-/**
- * ABI-encode Stargate withdrawal parameters
- */
-export function encodeStargateV2Payload({
-  layerZeroEndpointId,
-}: DecodedStargateV2Payload): EncodedStargateV2Payload {
-  return ethers.AbiCoder.defaultAbiCoder().encode(
-    ['uint32'],
-    [layerZeroEndpointId],
-  );
-}
-
-/**
- * Returns the encoded `bridgeAdapterPayload`
- */
-export function getEncodedWithdrawalPayloadForBridgeTarget(
-  bridgeTarget: BridgeTarget,
-  sandbox = false,
-): EncodedStargateV2Payload {
-  const targetConfig =
-    sandbox ?
-      StargateV2Config.testnet[bridgeTarget]
-    : StargateV2Config.mainnet[bridgeTarget];
-
-  if (!targetConfig || !targetConfig.isSupported) {
-    throw new Error(
-      `Stargate withdrawals not supported to chain ${targetConfig.target} ${sandbox ? 'testnet' : 'mainnet'} (Chain ID: ${String(targetConfig.evmChainId)})`,
-    );
-  }
-
-  return encodeStargateV2Payload({
-    layerZeroEndpointId: targetConfig.layerZeroEndpointId,
-  });
-}
+import { StargateV2Config } from './config';
 
 /**
  * Deposit funds cross-chain into the Exchange using Stargate
@@ -279,7 +143,7 @@ export async function estimateDepositFees(
   );
 }
 
-export async function estimateDepositViaForwarderFees(
+async function estimateDepositViaForwarderFees(
   sourceStargateTarget: StargateV2Target,
   parameters: {
     exchangeLayerZeroAdapterAddress?: string;
@@ -335,7 +199,7 @@ export async function estimateDepositViaForwarderFees(
  * Estimate native gas fee needed to deposit USDC cross-chain into the Exchange
  * from Berachain using the Kuma OFT pathway
  */
-export async function estimateDepositFromBerachainFees(
+async function estimateDepositFromBerachainFees(
   parameters: {
     exchangeLayerZeroAdapterAddress?: string;
     minimumWithdrawQuantityMultiplierInPips: bigint;
@@ -495,94 +359,4 @@ async function getDepositViaForwarderSendParamAndSourceConfig(
   };
 
   return { sendParam, sourceConfig };
-}
-
-/**
- * Estimate the quantity of tokens delivered to the target chain via Stargate v2 withdrawal
- */
-export async function estimateStargateV2WithdrawQuantity(
-  parameters: {
-    exchangeLayerZeroAdapterAddress: string;
-    stargateForwarderAddress: string;
-    payload: string;
-    quantityInDecimal: string;
-  },
-  xchainProvider: ethers.Provider,
-  berachainProvider: ethers.Provider,
-  sandbox: boolean,
-): Promise<{
-  estimatedWithdrawQuantityInDecimal: string;
-  minimumWithdrawQuantityInDecimal: string;
-  willSucceed: boolean;
-}> {
-  const { layerZeroEndpointId: targetEndpointId } = decodeStargateV2Payload(
-    parameters.payload,
-  );
-
-  const exchangeStargateAdapter = ExchangeLayerZeroAdapter_v2__factory.connect(
-    parameters.exchangeLayerZeroAdapterAddress,
-    xchainProvider,
-  );
-
-  const [
-    estimatedWithdrawQuantityInAssetUnits,
-    minimumWithdrawQuantityInAssetUnits,
-    poolDecimals,
-  ] = await exchangeStargateAdapter.estimateWithdrawQuantityInAssetUnits(
-    // Funds must always be withdrawn to Berachain first regardless of final target
-    (sandbox ? StargateV2Config.testnet : StargateV2Config.mainnet)[
-      StargateV2Target.STARGATE_BERACHAIN
-    ].layerZeroEndpointId,
-    decimalToPip(parameters.quantityInDecimal),
-  );
-  let estimatedWithdrawQuantityInDecimal = assetUnitsToDecimal(
-    estimatedWithdrawQuantityInAssetUnits,
-    Number(poolDecimals),
-  );
-  let minimumWithdrawQuantityInDecimal = assetUnitsToDecimal(
-    minimumWithdrawQuantityInAssetUnits,
-    Number(poolDecimals),
-  );
-
-  const target = stargateV2TargetForLayerZeroEndpointId(
-    targetEndpointId,
-    sandbox,
-  );
-  if (!target || !getStargateV2TargetConfig(target, sandbox)) {
-    throw new Error(`Unsupported Layerzero Endpoint ID ${targetEndpointId}`);
-  }
-  // If the final target is not Berachain, an bridge tx is needed via the forwarder
-  if (target !== StargateV2Target.STARGATE_BERACHAIN) {
-    const stargateForwarder = KumaStargateForwarder_v1__factory.connect(
-      parameters.stargateForwarderAddress,
-      berachainProvider,
-    );
-
-    const [
-      estimatedForwardedQuantityInAssetUnits,
-      minimumForwardedQuantityInAssetUnits,
-      forwardedPoolDecimals,
-    ] = await stargateForwarder.loadEstimatedForwardedQuantityInAssetUnits(
-      targetEndpointId,
-      decimalToPip(estimatedWithdrawQuantityInDecimal),
-    );
-    estimatedWithdrawQuantityInDecimal = assetUnitsToDecimal(
-      estimatedForwardedQuantityInAssetUnits,
-      Number(forwardedPoolDecimals),
-    );
-    minimumWithdrawQuantityInDecimal = assetUnitsToDecimal(
-      minimumForwardedQuantityInAssetUnits,
-      Number(forwardedPoolDecimals),
-    );
-  }
-
-  const willSucceed =
-    BigInt(estimatedWithdrawQuantityInAssetUnits) >=
-    BigInt(minimumWithdrawQuantityInAssetUnits);
-
-  return {
-    estimatedWithdrawQuantityInDecimal,
-    minimumWithdrawQuantityInDecimal,
-    willSucceed,
-  };
 }
