+++ src/bridge/index.ts
@@ -1,2 +1,4 @@
-export * from '#bridge/bridge';
 export * from '#bridge/config';
+export * from '#bridge/deposit';
+export * from '#bridge/utils';
+export * from '#bridge/withdraw';
